move costants into utils
