+++ src/Components/ConfigurationPane/FilterFormElements/ForksCountInput.tsx
@@ -1,5 +1,5 @@
 import { FormControl, TextInput } from "@primer/react";
-import { MIN_FORKS, MAX_FORKS } from "@Constants";
+import { MIN_FORKS, MAX_FORKS } from "@Utils/Constants";
 
 
 type ForksCountInputValidation = "LessThanMinForksError" | "GreaterThanMaxForksError";
